+++ tests/fixtures/join-jsonc/object/with-block-comment.in.ts
@@ -1,0 +1,5 @@
+{
+  /* which port */
+  "port": 5432,
+  "host": "localhost"
+}